Factored out sum function for Resources.
This is the start of a patch chain for converting `Resources` to `hashmap<SlaveID, Resources>` in places where we have an aggregate of resources across multiple slaves. In short, scalar resources aggregate correctly across multiple slaves, but non-scalar resources (e.g. ports) do not aggregate correctly. Please refer to the JIRA ticket for further details.

The function `sum` is factored out to facilitate small patches. The following is the strategy and why this function will be useful:

(1) Update a variable `resources` in a component (e.g. master) from `Resources` to `hashmap<SlaveID, Resources>`.
(2) For callsites where we pass `resources` to another component (e.g. allocator), we'll pass `sum(resources.values())` instead to preserve the existing behavior in the other component.

Review: https://reviews.apache.org/r/31183