Quota: Introduced capacity heuristic for quota set requests.
The heuristic tests whether the total quota, including the new request, does not exceed the sum of non-static cluster resources. The goal of this checkis to determine whether a user may accidentally request an amount of resources that would prevent frameworks without quota from getting any offers. A force flag will be introduced in the future and will allow users to bypass this check.

Review: https://reviews.apache.org/r/38110