Updated `TestContainerizer` to support default actions.
We use `ON_CALL` and `WillByDefault` here to specify the default
actions. This allows the tests to leverage the `DoDefault` action.
However, `ON_CALL` results in a "Uninteresting mock function call"
warning unless each test puts expectations in place. As a result,
we also use `EXPECT_CALL` and `WillRepeatedly` to get the best of
both worlds: the ability to use `DoDefault` and no warnings when
expectations are not explicit.

Review: https://reviews.apache.org/r/44993