Refactored command executor's shutdown logic.
Even though the command executor manages a single task, killTask() and
shutdown() differ in time period the executor has to finalize. If a
killTask is issued, the executor may use as much time as needed to
wait for the underlying task (specified in the kill policy) to exit.
If asked to shutdown, the executor is limited by the agent, that
destroys the container altogether after a certain timeout. Moreover,
in the latter case the user is usually less interested in graceful
shutdown, because the framework is being removed completely.