Updated command executor to support kill policy in `killTask()`.
Kill policy can be provided in a `killTask()` callback. In this case
it should take precedence over kill policy specified when the task
was launched. When `killTask()` is called multiple times during the
task termination, the signal escalation timeout (the time a task has
between SIGTERM and SIGKILL) is updated.