Updated the ZooKeeper patch.
This patch was generated with ZOOKEEPER-2841 and ZOOKEEPER-1643 (the
original patch). This adds a CMake build system to ZooKeeper and fixes
the Windows build. The exact patch lives at:

https://github.com/andschwa/zookeeper/tree/cmake-backport-3.4.8

Note that the deletion of the .sln file was removed in comparison to the
PR to backport these patches to ZooKeeper's 3.4 branch. This is because
the file is committed with CRLF line endings, but the generated patch
does not include them. This confuses GNU Patch, and the easiest solution
was to ignore the deletion of the file.

When PR #313 is accepted to ZooKeeper and the next patch version of 3.4
is released, we can replace the tarball and delete this patch.

Review: https://reviews.apache.org/r/61308