Fixed up 3rdparty CMake modules.
This removes `Mesos3rdpartyConfigure.cmake` entirely, as the only
necessary code in it was moved to `PatchCommand.cmake`, which was
consolidated from `libprocess/cmake/macros` to `cmake`.

`External.cmake` was similarly moved.

The `patch_cmd` function was fixed to work on Windows, allowing
`3rdparty/CMakeLists.txt` to be cleaned up further by consolidating the
setting of different patch commands.