Updated `CompilationConfigure.cmake` for imported libraries.
Windows and Linux have converged on several libraries (ZooKeeper,
Protobuf, and libevent) such that the same bundled version is used, so
the warning message was updated.

Third-party specific compilation flags were moved to the import of each
library, so this configuration no longer needs to include it.

The minimum CMake version was updated and usupports `string(TIMESTAMP)`,
deprecating the `date` work-around.