Added non-src files to CMake project to facilitate indexing.
Some IDEs and indexing tools build the project structure from CMake
target definitions. In this case files that are not essential for
building the target and are omitted from the target declaration,
do not appear in the project either and hence are not searchable
inside the IDE or indexing tool.

Review: https://reviews.apache.org/r/62071