Ensured scheduled events are handled before executor driver goes away.
This is similar to the fix for the scheduler driver in MESOS-4029.

This change adds the ability to stop running the executor library
process so that no future callbacks are delivered to the executor.
This helps, for example, during testing to ensure no further callbacks
happen to stack allocated mock objects.

This change also fixes the tests using the executor library by
explicitly invoking `stop()` from the testing interface. This ensures
that no further callbacks are delivered to the executor. For one-off
async callbacks that are already on the libprocess queue we need to do
a `Clock::settle` to ensure they are executed before the mock object
goes out of scope.