Updated libprocess routing to treat trailing '/' as insignificant.
Prior to this patch, adding a trailing '/' to a valid URL path, e.g.,
"/state/", yielded a 404 response. This patch ensures that two URLs
which differ only in trailing '/' produce the same result.

This is achieved by (1) trimming a trailing '/' when processing a
request and (2) disallowing `route()` calls from using `/`.

Review: https://reviews.apache.org/r/64574