Logged socket create/connect failures on warning level.
If a socket create or connect failure occurs during link() or send(),
the reason for error is not propagated to the library user. Also, the
data being sent or queued is silently dropped on the floor. The socket
code does not log itself on a higher level when an error situation
occurs. The only trace is the log entries touched in this patch:
having them at warning level will significantly simplify debugging.

This patch also consistently logs send / link target.

Review: https://reviews.apache.org/r/65048