Added MemoryProfiler class to Libprocess.
This class exposes profiling functionality of jemalloc memory allocator
when it is detected to be the memory allocator of the current process.

In particular, it gives developers an easy method to collect and
access heap profiles which report which pieces of code were
responsible for allocating memory.

Review: https://reviews.apache.org/r/63368/